removed Clock Trait from individual api's and moved it to common folder
