+++ audio-api/src/main/scala/no/ndla/audioapi/controller/AudioController.scala
@@ -27,7 +27,7 @@ import no.ndla.audioapi.model.api.{
 import no.ndla.audioapi.model.domain.{AudioType, SearchSettings}
 import no.ndla.audioapi.repository.AudioRepository
 import no.ndla.audioapi.service.search.{AudioSearchService, SearchConverterService}
-import no.ndla.audioapi.service.{Clock, ConverterService, ReadService, WriteService}
+import no.ndla.audioapi.service.{ConverterService, ReadService, WriteService}
 import no.ndla.language.Language
 import org.json4s.ext.EnumNameSerializer
 import org.json4s.{DefaultFormats, Formats}
@@ -46,7 +46,6 @@ trait AudioController {
     with AudioSearchService
     with Role
     with User
-    with Clock
     with SearchConverterService
     with ConverterService
     with NdlaController
