--- audio-api/src/main/scala/no/ndla/audioapi/service/Clock.scala
@@ -1,21 +1,0 @@
-/*
- * Part of NDLA audio-api
- * Copyright (C) 2017 NDLA
- *
- * See LICENSE
- */
-
-package no.ndla.audioapi.service
-
-import java.time.LocalDateTime
-
-trait Clock {
-  val clock: SystemClock
-
-  class SystemClock {
-
-    def now(): LocalDateTime = {
-      LocalDateTime.now()
-    }
-  }
-}
