+++ audio-api/src/main/scala/no/ndla/audioapi/service/WriteService.scala
@@ -15,6 +15,7 @@ import no.ndla.audioapi.model.{api, domain}
 import no.ndla.audioapi.model.domain.Audio
 import no.ndla.audioapi.repository.{AudioRepository, SeriesRepository}
 import no.ndla.audioapi.service.search.{AudioIndexService, SeriesIndexService, TagIndexService}
+import no.ndla.common.Clock
 import no.ndla.language.Language.findByLanguageOrBestEffort
 import org.scalatra.servlet.FileItem
