+++ common/src/main/scala/no/ndla/common/DateParser.scala
@@ -1,3 +1,11 @@
+/*
+ * Part of NDLA common.
+ * Copyright (C) 2022 NDLA
+ *
+ * See LICENSE
+ *
+ */
+
 package no.ndla.common
 
 import java.time.{LocalDateTime, ZoneOffset}
