--- concept-api/src/main/scala/no/ndla/conceptapi/service/Clock.scala
@@ -1,21 +1,0 @@
-/*
- * Part of NDLA concept-api.
- * Copyright (C) 2019 NDLA
- *
- * See LICENSE
- */
-
-package no.ndla.conceptapi.service
-
-import java.time.LocalDateTime
-
-trait Clock {
-  val clock: SystemClock
-
-  class SystemClock {
-
-    def now(): LocalDateTime = {
-      LocalDateTime.now()
-    }
-  }
-}
