--- draft-api/src/main/scala/no/ndla/draftapi/service/Clock.scala
@@ -1,21 +1,0 @@
-/*
- * Part of NDLA draft-api.
- * Copyright (C) 2017 NDLA
- *
- * See LICENSE
- */
-
-package no.ndla.draftapi.service
-
-import java.time.LocalDateTime
-
-trait Clock {
-  val clock: SystemClock
-
-  class SystemClock {
-
-    def now(): LocalDateTime = {
-      LocalDateTime.now()
-    }
-  }
-}
