+++ draft-api/src/test/scala/no/ndla/draftapi/TestEnvironment.scala
@@ -10,6 +10,7 @@ package no.ndla.draftapi
 import com.amazonaws.services.s3.AmazonS3
 import com.typesafe.scalalogging.LazyLogging
 import com.zaxxer.hikari.HikariDataSource
+import no.ndla.common.Clock
 import no.ndla.draftapi.auth.User
 import no.ndla.draftapi.caching.MemoizeHelpers
 import no.ndla.draftapi.controller._
