+++ image-api/src/main/scala/no/ndla/imageapi/ComponentRegistry.scala
@@ -10,6 +10,7 @@ package no.ndla.imageapi
 
 import com.amazonaws.regions.Regions
 import com.amazonaws.services.s3.{AmazonS3, AmazonS3ClientBuilder}
+import no.ndla.common.Clock
 import no.ndla.imageapi.auth.{Role, User}
 import no.ndla.imageapi.controller._
 import no.ndla.imageapi.integration._
