--- learningpath-api/src/main/scala/no/ndla/learningpathapi/service/Clock.scala
@@ -1,22 +1,0 @@
-/*
- * Part of NDLA learningpath-api.
- * Copyright (C) 2016 NDLA
- *
- * See LICENSE
- *
- */
-
-package no.ndla.learningpathapi.service
-
-import java.time.LocalDateTime
-
-trait Clock {
-  val clock: SystemClock
-
-  class SystemClock {
-
-    def now(): LocalDateTime = {
-      LocalDateTime.now()
-    }
-  }
-}
